+++ languages/de/tl_module.php
@@ -1,0 +1,35 @@
+<?php
+
+/**
+ * polls extension for Contao Open Source CMS
+ * 
+ * Copyright (C) 2013 Codefog
+ * 
+ * @package polls
+ * @link    http://codefog.pl
+ * @author  Kamil Kuzminski <kamil.kuzminski@codefog.pl>
+ * @author  Sebastian Seidel <sebastian@derastronaut.de>
+ * @license LGPL
+ */
+
+
+/**
+ * Fields
+ */
+$GLOBALS['TL_LANG']['tl_module']['poll_active']   = array('Aktive Artikel', 'Hier können Sie wählen, wie aktive Umfragen behandelt werden sollen.');
+$GLOBALS['TL_LANG']['tl_module']['poll_visible']  = array('Sichtbare Artikel', 'Hier können Sie wählen, wie sichtbare Umfragen behandelt werden sollen.');
+$GLOBALS['TL_LANG']['tl_module']['poll_featured'] = array('Hervorgehobene Artikel', 'Hier können Sie wählen, wie hervorgehobene Umfragen behandelt werde sollen.');
+
+
+/**
+ * Reference
+ */
+$GLOBALS['TL_LANG']['tl_module']['poll_active']['all']   = 'Alle Artikel';
+$GLOBALS['TL_LANG']['tl_module']['poll_active']['yes']   = 'Nur aktive Artikel';
+$GLOBALS['TL_LANG']['tl_module']['poll_active']['no']    = 'Nur inaktive Artikel';
+$GLOBALS['TL_LANG']['tl_module']['poll_visible']['all']  = 'Alle Artikel';
+$GLOBALS['TL_LANG']['tl_module']['poll_visible']['yes']  = 'Nur sichtbare Artikel';
+$GLOBALS['TL_LANG']['tl_module']['poll_visible']['no']   = 'Nur unsichtbare Artikel';
+$GLOBALS['TL_LANG']['tl_module']['poll_featured']['all'] = 'Alle Artikel';
+$GLOBALS['TL_LANG']['tl_module']['poll_featured']['yes'] = 'Nur hervorgehobene Artikel';
+$GLOBALS['TL_LANG']['tl_module']['poll_featured']['no']  = 'Nur nicht-hervorgehobene Artikel';