+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 polls Changelog
 ===============
 
+Version 1.0.1 stable (2013-05-11)
+---------------------------------
+
+### Improved
+- Added the German translation (thanks to Sebastian Seidel)
+
+### Improved
+- Added the "show results immedietaly" to behavior not voted
+
+### Fixed
+- Fixed the wrong poll behaviors
+
+
 Version 1.0.0 stable (2013-04-21)
 ---------------------------------